Fix the WebDAV URL discovery
Prior to this commit the WebDAV URL was constructed from the server hostname
(returned by the API) and the constant /uploads. It doesn't work longer,
because the API returns new WebDAV URL which is <hostname>/gdc/uploads. This fix
ensures the WebDAV URL is discovered from the API at runtime.